refactor(create-product-input): change double to big decimal in price

+++ infrastructure/src/main/java/com/kaua/ecommerce/infrastructure/product/models/CreateProductInput.java
@@ -2,10 +2,12 @@ package com.kaua.ecommerce.infrastructure.product.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record CreateProductInput(
         @JsonProperty("name") String name,
         @JsonProperty("description") String description,
-        @JsonProperty("price") double price,
+        @JsonProperty("price") BigDecimal price,
         @JsonProperty("quantity") int quantity,
         @JsonProperty("category_id") String categoryId,
         @JsonProperty("color_name") String colorName,
